[advanced] 유틸함수 parseDateTime, isOverlapping, findOverlappingEvents 분리
