+++ packages/assignment-7/src/hooks/useEvents.ts
@@ -2,7 +2,7 @@ import { useToast } from '@chakra-ui/react';
 import { useRef, useState } from 'react';
 import { Event, UseEventsReturn } from '../types';
 import { DUMMY_EVENTS } from '../dummys';
-import { getWeekDates } from '../utils';
+import { findOverlappingEvents, getWeekDates } from '../utils';
 import UseEventForm from './useEventForm';
 import UseNotification from './useNotificaiton';
 
@@ -205,29 +205,7 @@ function useEvents(): UseEventsReturn {
       notificationTime,
     };
 
-    // 날짜 문자열을 Date 객체로 변환하는 함수
-    const parseDateTime = (date: string, time: string): Date => {
-      return new Date(`${date}T${time}`);
-    };
-
-    // 두 일정이 겹치는지 확인하는 함수
-    const isOverlapping = (event1: Event, event2: Event): boolean => {
-      const start1 = parseDateTime(event1.date, event1.startTime);
-      const end1 = parseDateTime(event1.date, event1.endTime);
-      const start2 = parseDateTime(event2.date, event2.startTime);
-      const end2 = parseDateTime(event2.date, event2.endTime);
-
-      return start1 < end2 && start2 < end1;
-    };
-
-    // 겹치는 일정을 찾는 함수
-    const findOverlappingEvents = (newEvent: Event): Event[] => {
-      return events.filter(
-        (event) => event.id !== newEvent.id && isOverlapping(event, newEvent)
-      );
-    };
-
-    const overlapping = findOverlappingEvents(eventData);
+    const overlapping = findOverlappingEvents(eventData, events);
     if (overlapping.length > 0) {
       setOverlappingEvents(overlapping);
       setIsOverlapDialogOpen(true);
