+++ packages/assignment-7/src/utils.ts
@@ -1,4 +1,4 @@
-import { ResetFormSetters } from './types';
+import { Event, ResetFormSetters } from './types';
 
 export const resetFormState = (setters: ResetFormSetters) => {
   const [
@@ -58,4 +58,26 @@ export const formatMonth = (date: Date): string => {
   const year = date.getFullYear();
   const month = date.getMonth() + 1;
   return `${year}년 ${month}월`;
+};
+
+export const parseDateTime = (date: string, time: string): Date => {
+  return new Date(`${date}T${time}`);
+};
+
+export const isOverlapping = (event1: Event, event2: Event): boolean => {
+  const start1 = parseDateTime(event1.date, event1.startTime);
+  const end1 = parseDateTime(event1.date, event1.endTime);
+  const start2 = parseDateTime(event2.date, event2.startTime);
+  const end2 = parseDateTime(event2.date, event2.endTime);
+
+  return start1 < end2 && start2 < end1;
+};
+
+export const findOverlappingEvents = (
+  newEvent: Event,
+  events: Event[]
+): Event[] => {
+  return events.filter(
+    (event) => event.id !== newEvent.id && isOverlapping(event, newEvent)
+  );
 };
